Splitting out task like I did with time - glanceability
